+++ src/genui/main.py
@@ -191,6 +191,7 @@ class Window(
             inference_steps=self.steps_editor.value(),
             deepcache_enabled=self.deepcache_enabled_editor.isChecked(),
             use_karras_sigmas=self.karras_sigmas_editor.isChecked(),
+            use_vpred=self.vpred_editor.isChecked(),
         )
         # Send prompt to worker for start of generation.
         self.gen_worker.parent_conn.send(prompt)