+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Scheme of releases: `v<4-digit year>.<1 or 2-digit month>.<patch starting at 0>`
 
 ## [Unreleased]
 
+## [v2025.3.2] - 2025-03-18
+
 ### Added
 
 - Added support v-prediction models.
@@ -18,8 +20,6 @@ Scheme of releases: `v<4-digit year>.<1 or 2-digit month>.<patch starting at 0>`
 ### Fixed
 
 - Fixed error on generation after CUDA OutOfMemoryError. Clearing pipeline cache after OutOfMemoryError.
-
-### Changed
 
 ## [v2025.3.1] - 2025-03-13